Minor edits to remove warnings in conjunction with moving to Eclipse 4.2
git-svn-id: file:///home/jls/openjml/repo_master/OpenJML/trunk@2627 1627bae4-3225-46cb-a19f-684862636fd9

+++ OpenJML/runtime/org/jmlspecs/models/JMLChar.java
@@ -137,7 +137,7 @@ public /*@ pure @*/ class JMLChar implements JMLComparable {
       @    ensures \result >= 0 && (* \result is a hash code for theChar *);
       @*/
     public int hashCode() {
-        return ((int) value);
+        return value;
     }
   
     /** Compare this to op2, returning a comparison code.
